# packages/server/src/api/swagger/Api.v3.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: balanced-gym-app
  description: Balanced Gym App
tags:
  - name: Routine
  - name: Serie
  - name: Exercise
  - name: Muscle
  - name: MuscleGroup
  - name: Admin
paths:
  /routines:
    get:
      tags:
        - Routine
      description: Fetch all routines
      responses:
        "200":
          description: Returns all routines
  /routine/{routineId}:
    get:
      tags:
        - Routine
      description: Fetch the routine with the given Id
      parameters:
        - in: path
          name: routineId
          schema:
            type: string
          required: true
          example: 59f0c59d4e55c40d38868034
      responses:
        "200":
          description: Returns the routine with ID
  /routine/{routineId}/muscleGroup/{muscleGroupId}:
    get:
      tags:
        - Routine
      description: Fetch the muscleGroup with the given IDs
      parameters:
        - in: path
          name: routineId
          schema:
            type: string
          required: true
          example: 59f0c59d4e55c40d38868034
        - in: path
          name: muscleGroupId
          schema:
            type: string
          required: true
          example: 5dde9b568603d000157b627d
      responses:
        "200":
          description: Returns the routine with ID
  /routine/{routineId}/muscleGroup/{muscleGroupId}/target/{targetId}:
    get:
      tags:
        - Routine
      description: Fetch the target with the given IDs
      parameters:
        - in: path
          name: routineId
          schema:
            type: string
          required: true
          example: 59f0c59d4e55c40d38868034
        - in: path
          name: muscleGroupId
          schema:
            type: string
          required: true
          example: 5dde9b568603d000157b627d
        - in: path
          name: targetId
          schema:
            type: string
          required: true
          example: 5a55152ad9b873001457578f
      responses:
        "200":
          description: Returns the target with ID
  /updateSerie/{serieId}/exercise/{exerciseId}:
    patch:
      tags:
        - Serie
      description: Update the serie with the given Id
      parameters:
        - in: path
          name: serieId
          schema:
            type: string
          required: true
          example: 5a4ea837ae79ef0014fc737f
        - in: path
          name: exerciseId
          schema:
            type: string
          required: true
          example: 5a3eb620260b8700140e500e
      requestBody:
        description: the ISerieUpdate data
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reps:
                  type: number
                  example: 10
                weight:
                  type: number
                  example: 12.5
      responses:
        "200":
          description: Returns the exercise with the serie updated and hte updated serie
  /deleteSerie/{serieId}/exercise/{exerciseId}:
    delete:
      tags:
        - Serie
      description: Delete the serie with the given Id
      parameters:
        - in: path
          name: serieId
          schema:
            type: string
          required: true
        - in: path
          name: exerciseId
          schema:
            type: string
          required: true
          example: 5a3eb620260b8700140e500e
      responses:
        "200":
          description: Returns the exercise with the serie deleted and the deleted serie
  /newSerie/{exerciseId}:
    post:
      tags:
        - Serie
      description: Adds a new serie to the exercise with the given Id
      parameters:
        - in: path
          name: exerciseId
          schema:
            type: string
          required: true
          example: 5a3eb620260b8700140e500e
      requestBody:
        description: the suggestedSerie 
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reps:
                  type: number
                  example: 10
                weight:
                  type: number
                  example: 12.5
      responses:
        "200":
          description: Returns the exercise with the new serie and the new serie
  /muscles:
    get:
      tags:
        - Muscle
      description: Fetch all muscles (sorted by name)
      responses:
        "200":
          description: Returns all muscles
  /newMuscle:
    post:
      tags:
        - Admin
      description: Create a new muscle (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Biceps Brachii
                muscleURL:
                  type: string
                  example: BicepsInsideSmall.png
      responses:
        "201":
          description: Returns the created muscle
        "403":
          description: Admin access required
  /muscle/{muscleId}:
    patch:
      tags:
        - Admin
      description: Update a muscle (admin only)
      parameters:
        - in: path
          name: muscleId
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                muscleURL:
                  type: string
      responses:
        "200":
          description: Returns the updated muscle
        "403":
          description: Admin access required
  /muscleGroups:
    get:
      tags:
        - MuscleGroup
      description: Fetch all muscle groups (sorted by order)
      responses:
        "200":
          description: Returns all muscle groups
  /newExercise/{routineId}:
    post:
      tags:
        - Admin
      description: Create a new exercise in a routine (admin only)
      parameters:
        - in: path
          name: routineId
          schema:
            type: string
          required: true
          example: 59f0c59d4e55c40d38868034
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Barbell Bench Press
                muscleGroup:
                  type: string
                  example: 5dde9b568603d000157b627d
                target:
                  type: string
                  example: 5a55152ad9b873001457578f
                gifURL:
                  type: string
                  example: BBBenchPress.gif
                equipment:
                  type: string
                  example: Barbell Long
                synergists:
                  type: array
                  items:
                    type: string
                stabilizers:
                  type: array
                  items:
                    type: string
      responses:
        "201":
          description: Returns the created exercise
        "403":
          description: Admin access required
  /exercise/{exerciseId}:
    get:
      tags:
        - Exercise
      description: Fetch the exercise with the given Id
      parameters:
        - in: path
          name: exerciseId
          schema:
            type: string
          required: true
          example: 5a04c3471e403f31c973d699
      responses:
        "200":
          description: Returns the exercise with ID
    patch:
      tags:
        - Admin
      description: Update an exercise (admin only)
      parameters:
        - in: path
          name: exerciseId
          schema:
            type: string
          required: true
          example: 5a04c3471e403f31c973d699
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                muscleGroup:
                  type: string
                target:
                  type: string
                gifURL:
                  type: string
                equipment:
                  type: string
                synergists:
                  type: array
                  items:
                    type: string
                stabilizers:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Returns the updated exercise
        "403":
          description: Admin access required
    delete:
      tags:
        - Admin
      description: Delete an exercise and its series (admin only)
      parameters:
        - in: path
          name: exerciseId
          schema:
            type: string
          required: true
          example: 5a04c3471e403f31c973d699
      responses:
        "204":
          description: Exercise deleted successfully
        "403":
          description: Admin access required
  /spec:
    get:
      tags:
        - Specification
      responses:
        "200":
          description: Return the API specification
servers:
  - url: /api/v1